requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.2
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
joblib==1.4.2
Markdown==3.6
numpy==2.0.0
openai==1.35.7
proto-plus==1.24.0
protobuf==4.25.3
psycopg==3.1.19
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
scikit-learn==1.5.0
scipy==1.14.0
sniffio==1.3.1
sqlparse==0.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
youtube-transcript-api==0.6.2
